# api/openapi.yaml
openapi: 3.1.0
info:
  title: KOEN Internal Weather API
  version: 1.0.0
servers:
  - url: /
security: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
    Envelope:
      type: object
      required: [apiVersion, generatedAt, data]
      properties:
        apiVersion: {const: v1}
        generatedAt: {type: string, format: date-time}
        data: {}
    Batch:
      type: object
      additionalProperties: false
      required: [schemaVersion, batchId, source, createdAt, records, assets]
      properties:
        schemaVersion: {const: '1.0'}
        batchId: {type: string, pattern: '^[A-Z0-9][A-Z0-9._-]{7,127}$'}
        source: {type: string, minLength: 1}
        createdAt: {type: string, format: date-time}
        records: {type: object}
        assets: {type: array, maxItems: 20}
  responses:
    Envelope200:
      description: Versioned response envelope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Envelope'}
    Unauthorized:
      description: Bearer token required
paths:
  /health/live:
    get:
      operationId: getLiveness
      summary: Process liveness
      responses:
        '200': {description: Live}
  /health/ready:
    get:
      operationId: getReadiness
      summary: Configuration, database, and storage readiness
      responses:
        '200': {description: Ready}
        '503': {description: Not ready}
  /api/v1/ingest/batches:
    post:
      operationId: ingestBatch
      summary: Atomically ingest one schema 1.0 batch
      security: [{bearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Batch'}
      responses:
        '200': {description: Processed or idempotent duplicate}
        '400': {description: Validation failure}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '413': {description: Request too large}
  /api/v1/sites:
    get:
      operationId: listSites
      summary: List enabled sites
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/observations/latest:
    get:
      operationId: listLatestObservations
      summary: List the latest observation per site
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/observations:
    get:
      operationId: listObservations
      summary: List observation history
      security: [{bearer: []}]
      parameters:
        - {name: since, in: query, schema: {type: string, format: date-time}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500}}
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '400': {description: Invalid query}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/forecasts/latest:
    get:
      operationId: listLatestForecasts
      summary: List only the newest issue for each site and valid time
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/warnings/active:
    get:
      operationId: listActiveWarnings
      summary: List active warnings
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/radar/latest:
    get:
      operationId: getLatestRadar
      summary: Get the latest radar frame
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/radar/frames:
    get:
      operationId: listRadarFrames
      summary: List radar frames
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/typhoons/active:
    get:
      operationId: listActiveTyphoons
      summary: List active typhoons and forecast points
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/display/dashboard:
    get:
      operationId: getDashboard
      summary: Get the display dashboard aggregate
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/status:
    get:
      operationId: getIntegrationStatus
      summary: Get per-collector integration freshness
      security: [{bearer: []}]
      responses:
        '200': {$ref: '#/components/responses/Envelope200'}
        '401': {$ref: '#/components/responses/Unauthorized'}